+++ index.d.ts
@@ -18,6 +18,7 @@
 import Component from './src/component.js'
 import Application from './src/application.js'
 import Launch from './src/launch.js'
+import Plugin from './src/plugin.js'
 
 /**
  * Blits - The Lightning 3 App Development Framework
@@ -26,6 +27,7 @@ declare module Blits {
   export { Component }
   export { Application }
   export { Launch }
+  export { Plugin }
 }
 
 export default Blits