+++ index.js
@@ -18,9 +18,11 @@
 import Component from './src/component.js'
 import Application from './src/application.js'
 import Launch from './src/launch.js'
+import Plugin from './src/plugin.js'
 
 export default {
   Component,
   Application,
   Launch,
+  Plugin,
 }